+++ src/main/java/org/socialcars/sinziana/simulation/environment/jung/IEdge.java
@@ -36,7 +36,7 @@ public interface IEdge
      * length
      * @return length
      */
-    int length();
+    Double length();
 
     /**
      * orientation
